update migration, seeder, admin controller, mentee controller,view

+++ app/Http/Controllers/AdminController.php
@@ -95,12 +95,13 @@ class AdminController extends Controller
 
     //-------------------------------------------Mentor-------------------------------------------
 
-    // Mentor Function 
+    // Mentor Function
     public function mentor()
     {
         $data_mentor = Mentor::all();
         $totalMentor = Mentor::count();
-        return view('admin.mentor', compact(['data_mentor', 'totalMentor']));
+        $total_kelompok = Kelompok::count();
+        return view('admin.mentor', compact(['data_mentor', 'totalMentor', 'total_kelompok']));
     }
 
     // Get Mentor By ID
@@ -130,7 +131,7 @@ class AdminController extends Controller
             "alamat_mentor" => $request->alamat_mentor,
             "notelp_mentor" => $request->notelp_mentor,
             "status_mentor" => $request->status_mentor,
-            "slug" => Str::slug($request->nama_mentor, '-')
+            "slug" => Str::slug($userID.'-'.$request->nama_mentor, '-')
         ]);
 
         return redirect('/admin/mentor')->with('success', 'Mentor Berhasil ditambahkan !');
@@ -141,8 +142,15 @@ class AdminController extends Controller
     {
         $data_mentor = Mentor::where('slug', $slug)->get()->first();
         $dataUser = User::where('id', $data_mentor->user_id)->get()->first();
-        $dataUser->delete($dataUser);
-        return redirect('/admin/mentor')->with('success', 'Mentor Berhasil dihapus !');
+        $data_kelompok = Kelompok::where('mentor_id', $data_mentor->id)->get()->first();
+        $data_pertemuan = Pertemuan::where('mentor_id', $data_mentor->id)->get()->first();
+
+        if($data_kelompok != null && $data_pertemuan != null){
+            $dataUser->delete($dataUser);
+            return redirect('/admin/mentor')->with('success', 'Mentor Berhasil dihapus !');
+        }else{
+            return redirect('/admin/mentor')->with('success', 'Mentor Gagal dihapus, karena masih ada relasi !');
+        }
     }
 
     // Import Mentor
@@ -174,49 +182,105 @@ class AdminController extends Controller
     {
         // Masih Gabisa :v hehe
     }
-     //-------------------------------------------Mentee-------------------------------------------
+
+    //-------------------------------------------Mentee-------------------------------------------
     // Mentee Function
     public function mentee()
     {
         $data_mentee = Mentee::all();
         $totalMentee = Mentee::count();
+        $total_kelompok = Kelompok::count();
+        $total_kelas = Kelas::count();
         $data_angkatan = Angkatan::all();
         $data_jurusan = Jurusan::all();
-        $data_prodi = Prodi::all();
         $data_kelompok = Kelompok::all();
-        $data_kelas = Kelas::with('prodi', 'prodi.jurusan')->get();
-        return view('admin.mentee', compact('data_mentee', 'data_jurusan', 'data_prodi', 'data_kelas', 'data_kelompok','data_angkatan','totalMentee'));
+        return view('admin.mentee', compact(['data_mentee', 'data_jurusan', 'data_kelompok','data_angkatan','totalMentee', 'total_kelompok', 'total_kelas']));
     }
     // Detail Mentee
-    public function detailMentee($id_mentee){
-        $data_mentee = Mentee::find($id_mentee);
-        return view('admin.mentee.detail',compact('data_mentee'));
+    public function detailMentee($slug){
+        $data_mentee = Mentee::where('slug', $slug)->get()->first();
+        $data_angkatan = Angkatan::all();
+        $data_jurusan = Jurusan::all();
+        $data_kelompok = Kelompok::all();
+        return view('admin.mentee.detail',compact(['data_mentee', 'data_jurusan', 'data_kelompok','data_angkatan']));
     }
     //Add Mentee
     public function addMentee(Request $request)
     {
-
         $user = User::create([
             "role" => "Mentee",
             "name" => $request->nama_mentee,
-            "nim" => $request->nim_mentee,
-            "password" => Hash::make('mentee123'),
+            "email" => $request->nim_mentee."@student.polinema.ac.id",
+            "password" => Hash::make($request->nim_mentee),
         ]);
 
         $userID = DB::getPdo()->lastInsertId();
 
-        $mentee = Mentor::create([
+        $mentee = Mentee::create([
             "user_id" => $userID,
-            "nama_mentee" => $request->nama_mentor,
-            "nim_mentee" => $request->nim_mentor,
-            "alamat_mentor" => $request->alamat_mentor,
-            "notelp_mentor" => $request->notelp_mentor,
-            "status_mentor" => $request->status_mentor,
-            "slug" => Str::slug($request->nama_mentor, '-')
+            "nim_mentee" => $request->nim_mentee,
+            "nama_mentee" => $request->nama_mentee,
+            "kelas_id" => $request->kelas_id,
+            "kelompok_id" => $request->kelompok_id,
+            "angkatan_id" => $request->angkatan_id,
+            "status_mentee" => "tidak lulus",
+            "slug" => Str::slug($userID.'-'.$request->nama_mentee, '-')
         ]);
 
-        return redirect('/admin/mentor')->with('success', 'Mentor Berhasil ditambahkan !');
+        return redirect('/admin/mentee')->with('success', 'Mentee Berhasil ditambahkan !');
     }
+
+    public function getProdiByIdJurusan($id){
+        $prodi = Prodi::where('jurusan_id', $id)->get();
+        return $prodi;
+    }
+
+    public function getKelasByIdProdi($id){
+        $kelas = Kelas::where('prodi_id', $id)->get();
+        return $kelas;
+    }
+
+    // Get Mentor By ID
+    public function menteeById($id)
+    {
+        $data_mentor = Mentor::find($id);
+        return view('admin.mentor', ['data_mentor' => $data_mentor]);
+    }
+
+    // Delete Mentor
+    public function delMentee($slug)
+    {
+        $data_mentee = Mentee::where('slug', $slug)->get()->first();
+        $dataUser = User::where('id', $data_mentee->user_id)->get()->first();
+
+            $dataUser->delete($dataUser);
+            return redirect('/admin/mentee')->with('success', 'Mentee Berhasil dihapus !');
+    }
+
+    // Import Mentor
+    public function impMentee(Request $request)
+    {
+        Excel::import(new MentorImport,$request->file('data_mentor'));
+        return redirect('/admin/mentor')->with('success', 'Mentor Berhasil di Import !');
+    }
+
+    // Update Mentor
+    public function updMentee(Request $request,$id_mentee){
+        $data_mentee = Mentee::find($id_mentee);
+        $data_mentee->update($request->all());
+        return redirect('/admin/mentor')->with('success', 'Mentor Berhasil di Update !');;
+    }
+    // Export Mentor
+    public function exportMenteeExcel()
+    {
+        return Excel::download(new MentorExport, 'Mentor.xlsx');
+    }
+    // Export PDF
+    public function exportMentee()
+    {
+        // Masih Gabisa :v hehe
+    }
+
 
     //-------------------------------------------Users-------------------------------------------
 
@@ -262,7 +326,7 @@ class AdminController extends Controller
         return redirect('/admin/data')->with('success', 'Data Kelas Berhasil ditambahkan !');
     }
     // Delete Data
-    
+
         // Delete Jurusan
 
         // Delete Prodi
@@ -273,7 +337,7 @@ class AdminController extends Controller
 
         // Get Function
         public function materi()
-        {   
+        {
             $data_materi = Materi::all();
             $data_kegiatan = Kegiatan::all('minggu_kegiatan');
             $totalMateri = Materi::count();
@@ -284,9 +348,9 @@ class AdminController extends Controller
         {
             // $materi = Materi::create($request->all());
 
-            // Convert link materi untuk ditampilkan di embed 
+            // Convert link materi untuk ditampilkan di embed
 
-            // Dari input link youtube, dipecah berdasarkan / 
+            // Dari input link youtube, dipecah berdasarkan /
             //explode itu string ke array
             $last = explode("/",$request->link_materi);
 
@@ -380,7 +444,7 @@ class AdminController extends Controller
         $data_keluhan->update($request->all());
         return redirect('/admin/keluhan/')->with('success', 'Keluhan Berhasil dijawab !');
     }
-   
+
     //-------------------------------------------Pertemuan-------------------------------------------
 
     // Pertemuan Function
@@ -409,7 +473,7 @@ class AdminController extends Controller
             $data_pertemuan = Pertemuan::find($id_pertemuan);
             $data_pertemuan->delete($data_pertemuan);
             return redirect('/admin/pertemuan')->with('success', 'Pertemuan Berhasil dihapus !');
-        }        
+        }
         // Search Pertemuan
         public function cariPertemuan($id_pertemuan)
         {
@@ -454,6 +518,6 @@ class AdminController extends Controller
         $data_pengumuman = Pengumuman::find($id_pengumuman);
         $data_pengumuman->update($request->all());
         return redirect('/admin/pengumuman')->with('success', 'Pengumuman Berhasil di Update !');
-        
+
     }
 }
